refactor: merge JdbcTemplate and implement ArgumentPreparedStatementSetter
