+++ nextstep-jdbc/src/main/java/nextstep/jdbc/template/PreparedStatementSetter.java
@@ -3,6 +3,7 @@ package nextstep.jdbc.template;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 
+@FunctionalInterface
 public interface PreparedStatementSetter {
     void setValues(PreparedStatement pstmt) throws SQLException;
 }
