+++ nextstep-jdbc/src/main/java/nextstep/jdbc/template/RowMapper.java
@@ -3,6 +3,7 @@ package nextstep.jdbc.template;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
+@FunctionalInterface
 public interface RowMapper<T> {
     T mapRow(ResultSet resultSet) throws SQLException;
 }
